Progress on new merge and GUI updates
added new image elements for the search pane (such as btn images and
special images for searching/no results/ invalid search). Made a lot of
progress on the current merge to 'sections', started making a new
calender element to keep track of the current showing sections. Still
need to update the course banners (on th eleft) when courses are overlaping.